third_party/WebKit/LayoutTests/fast/canvas/canvas-font-consistency.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {

    var canvas = document.createElement('canvas');
    var ctx = canvas.getContext('2d');
    var fontNames = ["bold 13px Arial",
                    "italic 13px Arial",
                    "small-caps 13px Arial"];

    for (var i = 0; i < fontNames.length; i++) {
        var fontName = fontNames[i];
        ctx.font = fontName;
        assert_equals(ctx.font, fontName);
    }
    
}, 'Series of tests to ensure that context.font is consistent.');
</script>
</body>
</html>
